+++ src/main/java/com/contentful/java/cda/LocalizedResource.java
@@ -6,7 +6,43 @@ import java.util.Map;
  * Represents a resource which may contain field values for multiple locales.
  */
 public abstract class LocalizedResource extends CDAResource {
-  String locale;
+  public class Localizer {
+    private final String locale;
+
+    Localizer(String locale) {
+      this.locale = locale;
+    }
+
+    /**
+     * Extracts a field from the fields set of the active locale, result type is inferred.
+     *
+     * @param key field key.
+     * @param <T> type.
+     * @return field value, null if it doesn't exist.
+     */
+    @SuppressWarnings("unchecked")
+    public <T> T getField(String key) {
+      final Map<String, T> value = (Map<String, T>) fields.get(key);
+      if (value == null) {
+        return null;
+      }
+
+      return getFieldForFallbackLocale(value, locale);
+    }
+
+    <T> T getFieldForFallbackLocale(Map<String, T> value, String locale) {
+      if (locale == null) {
+        return null;
+      }
+
+      final T localized = value.get(locale);
+      if (localized != null) {
+        return localized;
+      } else {
+        return getFieldForFallbackLocale(value, fallbackLocaleMap.get(locale));
+      }
+    }
+  }
 
   String defaultLocale;
 
@@ -17,33 +53,36 @@ public abstract class LocalizedResource extends CDAResource {
   Map<String, Object> rawFields;
 
   /**
+   * Creates an object to be used for returning different field in one locale.
+   *
+   * @param locale pointing to a locale in the environment.
+   * @return localizer to localize the fields.
+   */
+  public Localizer localize(String locale) {
+    return new Localizer(locale);
+  }
+
+  /**
+   * Get a field using the environments default locale.
+   *
+   * @param key field key.
+   * @param <T> type.
+   * @return field value, null if it doesn't exist.
+   * @see #localize(String)
+   */
+  public <T> T getField(String key) {
+    return localize(defaultLocale).getField(key);
+  }
+
+  /**
    * Extracts a field from the fields set of the active locale, result type is inferred.
    *
    * @param key field key.
    * @param <T> type.
    * @return field value, null if it doesn't exist.
    */
-  @SuppressWarnings("unchecked")
-  public <T> T getField(String key) {
-    final Map<String, T> value = (Map<String, T>) fields.get(key);
-    if (value == null) {
-      return null;
-    }
-
-    return getFieldForFallbackLocale(value, locale);
-  }
-
-  private <T> T getFieldForFallbackLocale(Map<String, T> value, String locale) {
-    if (locale == null) {
-      return null;
-    }
-
-    final T localized = value.get(locale);
-    if (localized != null) {
-      return localized;
-    } else {
-      return getFieldForFallbackLocale(value, fallbackLocaleMap.get(locale));
-    }
+  public <T> T getField(String locale, String key) {
+    return localize(locale).getField(key);
   }
 
   /**
@@ -51,37 +90,5 @@ public abstract class LocalizedResource extends CDAResource {
    */
   public Map<String, Object> rawFields() {
     return rawFields;
-  }
-
-  /**
-   * @return returns the active locale code for this resource.
-   */
-  public String locale() {
-    return locale;
-  }
-
-  /**
-   * Switches the locale to the one matching the given locale code.
-   *
-   * @param locale the locale to be set.
-   */
-  public void setLocale(String locale) {
-    this.locale = locale;
-  }
-
-  void setFallbackLocaleMap(Map<String, String> fallbackLocaleMap) {
-    this.fallbackLocaleMap = fallbackLocaleMap;
-  }
-
-  Map<String, String> fallbackLocaleMap() {
-    return fallbackLocaleMap;
-  }
-
-  void setDefaultLocale(String defaultLocale) {
-    this.defaultLocale = defaultLocale;
-  }
-
-  String defaultLocale() {
-    return defaultLocale;
   }
 }