+++ CHANGELOG.md
@@ -4,8 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Version [10.0.0] - TBD
 - Add `locales` endpoint
+- Add `Localizer` to make localizing entries thread save.
 - Update cache to store locales and not spaces
 - Remove locales from `Space`
+- Remove `setLocale` on resources (entries and assets). Use `localize()` instead.
 
 ## Version [9.1.1] - (2018-03-21)
 - Fix: numbers in http headers default back to english.